dev: Use minimal set of SYSFS flies for intelligent CAN controllers
Active or intelligent CAN controllers do the setting of the bit-timing
internally rendering most of the standard bit-timing parameters for
the passive CAN controllers unnecessary. Usually just an index to a
fixed bit-time setting can be selected. For this purpose, just the
SYSFS file "can_bittiming/bitrate" will be created, if the CAN
bit-timing constants are not provided for the device.

Signed-off-by: Wolfgang Grandegger <wg@grandegger.com>