can: mscan: fix TX message handling to ensure proper message ordering
An out-of-order TX message problem has been reported and demonstrated
by Vladislav. The TX message buffers were not handled properly if
just one message was pending. This patch remembers the id of the buffer
used previously and increase the buffer priority if the next one has a
lower id number. This allows up to approx. 3*256 messages to be sent
until the lowest buffer priority of 255 is reached.

Signed-off-by: Wolfgang Grandegger <wg@grandegger.com>