softing_cs: fix for net-next-2.6 (2.6.33+)
The pcmcia_request_window() is postponed a little, as some flags must be
set before this call (I believe). It is a small change, but with some
noise in the patch. Same with softing_card_irq().

The netif_start_queue() is omitted, as softing_cycle() does a
netif_wake_queue().

Signed-off-by: Kurt Van Dijck <kurt.van.dijck@eia.be>
Signed-off-by: Oliver Hartkopp <oliver@hartkopp.net>