cangw: Do not use skb->sk to detect already routed CAN frames.
As the latest changes to the can-gw have shown, the use of skb->sk created
several problems and still looks like a bad hack.

While checking the struct skbuff for usable containers to detect routed CAN-
frames that do not interfere with other sophisticated network technique, the
pointers to the [transport|network|mac]_header looked interesting.

So we mark routed frames by setting some mac header length which is not
relevant for the CAN frames located in the skb->data section.

As dev->header_ops is not set in CAN netdevices no one is ever accessing the
various header offsets in the CAN skbuffs anyway. E.g. using the packet socket
to read CAN frames is still working after this change to gw.c .