isotp: correct padding policy for flow control (FC) PDU
When defining a TX padding for an isotp socket this padding has to be valid for
every PDU (SF, FC, last CF) which is send from this node. Due to a logfile
which was taken from a real world case it looked like the FC has to have the
rx padding value for tx which is not the general case (and inconsistent).

Correct the padding policy in a way that the padding for a specific direction
e.g. TX is used for all paddable PDU types send from this communication entity.
This unlikely might break existing applications using the content checking of
padding bytes. But I assume this not to be a real world problem as there would
have been complains about the former implementation if anyone had ever used it.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>