isotp: add option to check optimized frame length
Additionally to checking the correct padding in content and size the
functionality to check the correct (minimal) length is implemented.

When setting the CAN_ISOTP_CHK_PAD_LEN flag without giving the padding value
(CAN_ISOTP_RX_PADDING is not set) the frame length is ckecked for the
minimum possible ('optimized') length. This check also takes care of the
mandatory padding which comes with CAN FD.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>