net: pass info struct via netdevice notifier
The commit 351638e7deeed2ec ("net: pass info struct via netdevice notifier")
from Jiri Pirko was commited on 2013/05/28 for Linux 3.11

But the introduced API was first used with Linux 4.9 - which caused segfaults
with Linux 4.9-rc kernels. This patch introduces the orignal changes in
the netdevice notifier call for the notifier users.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>